walasz/ftc2/requirements.txt
django-rest-auth
django
djangorestframework
django-allauth
faker
pytest